pyproject.toml
[project]
name = "nonebot_plugin_group_config"
version = "0.1.0"
description = "Nonebot2 群聊配置信息存储与管理插件"
readme = "README.md"
dependencies = [
    "nonebot2>=2.3.0",
    "nonebot-plugin-uninfo>=0.6.0",
]
requires-python = ">=3.9, <4.0"

[tool.nonebot]
plugins = ["nonebot_plugin_group_config"]
